@@ -92,7 +92,7 @@ module SPARQL
     #
     #   Defaults  `User-Agent` header, unless one is specified.
     # @option options [Hash] :read_timeout
-    def initialize(url, options = {}, &block)
+    def initialize(url, **options, &block)
       case url
       when RDF::Queryable
         @url, @options = url, options.dup
@@ -128,8 +128,8 @@ module SPARQL
     #
     # @param (see Query.ask)
     # @return [Query]
-    def ask(*args)
-      call_query_method(:ask, *args)
+    def ask(*args, **options)
+      call_query_method(:ask, *args, **options)
     end
 
     ##
@@ -137,8 +137,8 @@ module SPARQL
     #
     # @param (see Query.select)
     # @return [Query]
-    def select(*args)
-      call_query_method(:select, *args)
+    def select(*args, **options)
+      call_query_method(:select, *args, **options)
     end
 
     ##
@@ -146,8 +146,8 @@ module SPARQL
     #
     # @param (see Query.describe)
     # @return [Query]
-    def describe(*args)
-      call_query_method(:describe, *args)
+    def describe(*args, **options)
+      call_query_method(:describe, *args, **options)
     end
 
     ##
@@ -155,8 +155,8 @@ module SPARQL
     #
     # @param (see Query.construct)
     # @return [Query]
-    def construct(*args)
-      call_query_method(:construct, *args)
+    def construct(*args, **options)
+      call_query_method(:construct, *args, **options)
     end
 
     ##
@@ -179,15 +179,15 @@ module SPARQL
     #   client.insert_data(data)
     #
     # @example Inserting data into a named graph
-    #   client.insert_data(data, :graph => "http://example.org/")
+    #   client.insert_data(data, graph: "http://example.org/")
     #
     # @param  [RDF::Enumerable] data
     # @param  [Hash{Symbol => Object}] options
     # @option options [RDF::URI, String] :graph
     # @return [void] `self`
     # @see    http://www.w3.org/TR/sparql11-update/#insertData
-    def insert_data(data, options = {})
-      self.update(Update::InsertData.new(data, options))
+    def insert_data(data, **options)
+      self.update(Update::InsertData.new(data, **options))
     end
 
     ##
@@ -200,15 +200,15 @@ module SPARQL
     #   client.delete_data(data)
     #
     # @example Deleting data from a named graph
-    #   client.delete_data(data, :graph => "http://example.org/")
+    #   client.delete_data(data, graph: "http://example.org/")
     #
     # @param  [RDF::Enumerable] data
     # @param  [Hash{Symbol => Object}] options
     # @option options [RDF::URI, String] :graph
     # @return [void] `self`
     # @see    http://www.w3.org/TR/sparql11-update/#deleteData
-    def delete_data(data, options = {})
-      self.update(Update::DeleteData.new(data, options))
+    def delete_data(data, **options)
+      self.update(Update::DeleteData.new(data, **options))
     end
 
     ##
@@ -223,8 +223,8 @@ module SPARQL
     # @option options [RDF::URI, String] :graph
     # @return [void] `self`
     # @see    http://www.w3.org/TR/sparql11-update/#deleteInsert
-    def delete_insert(delete_graph, insert_graph = nil, where_graph = nil, options = {})
-      self.update(Update::DeleteInsert.new(delete_graph, insert_graph, where_graph, options))
+    def delete_insert(delete_graph, insert_graph = nil, where_graph = nil, **options)
+      self.update(Update::DeleteInsert.new(delete_graph, insert_graph, where_graph, **options))
     end
 
     ##
@@ -240,8 +240,8 @@ module SPARQL
     # @option options [Boolean] :silent
     # @return [void] `self`
     # @see    http://www.w3.org/TR/sparql11-update/#clear
-    def clear_graph(graph_uri, options = {})
-      self.clear(:graph, graph_uri, options)
+    def clear_graph(graph_uri, **options)
+      self.clear(:graph, graph_uri, **options)
     end
 
     ##
@@ -267,7 +267,7 @@ module SPARQL
     #   @option options [Boolean] :silent
     #   @return [void] `self`
     #
-    # @overload clear(what, *arguments, options = {})
+    # @overload clear(what, *arguments, **options)
     #   @param  [Symbol, #to_sym] what
     #   @param  [Array] arguments splat of other arguments to {Update::Clear}.
     #   @param  [Hash{Symbol => Object}] options
@@ -281,9 +281,9 @@ module SPARQL
 
     ##
     # @private
-    def call_query_method(meth, *args)
+    def call_query_method(meth, *args, **options)
       client = self
-      result = Query.send(meth, *args)
+      result = Query.send(meth, *args, **options)
       (class << result; self; end).send(:define_method, :execute) do
         client.query(self)
       end
@@ -308,20 +308,20 @@ module SPARQL
     # @return [Array<RDF::Query::Solution>]
     # @raise [IOError] if connection is closed
     # @see    http://www.w3.org/TR/sparql11-protocol/#query-operation
-    def query(query, options = {})
+    def query(query, **options)
       @op = :query
       @alt_endpoint = options[:endpoint]
       case @url
       when RDF::Queryable
         require 'sparql' unless defined?(::SPARQL::Grammar)
         begin
-          SPARQL.execute(query, @url, options)
+          SPARQL.execute(query, @url, **options)
         rescue SPARQL::MalformedQuery
           $stderr.puts "error running #{query}: #{$!}"
           raise
         end
       else
-        parse_response(response(query, options), options)
+        parse_response(response(query, **options), **options)
       end
     end
 
@@ -336,15 +336,15 @@ module SPARQL
     # @return [void] `self`
     # @raise [IOError] if connection is closed
     # @see    http://www.w3.org/TR/sparql11-protocol/#update-operation
-    def update(query, options = {})
+    def update(query, **options)
       @op = :update
       @alt_endpoint = options[:endpoint]
       case @url
       when RDF::Queryable
         require 'sparql' unless defined?(::SPARQL::Grammar)
-        SPARQL.execute(query, @url, options.merge(update: true))
+        SPARQL.execute(query, @url, update: true, **options)
       else
-        response(query, options)
+        response(query, **options)
       end
       self
     end
@@ -359,7 +359,7 @@ module SPARQL
     # @option options [Hash] :headers
     # @return [String]
     # @raise [IOError] if connection is closed
-    def response(query, options = {})
+    def response(query, **options)
       headers = options[:headers] || @headers
       headers['Accept'] = options[:content_type] if options[:content_type]
       request(query, headers) do |response|
@@ -380,7 +380,7 @@ module SPARQL
     # @param  [Net::HTTPSuccess] response
     # @param  [Hash{Symbol => Object}] options
     # @return [Object]
-    def parse_response(response, options = {})
+    def parse_response(response, **options)
       case options[:content_type] || response.content_type
         when NilClass
           response.body
@@ -395,7 +395,7 @@ module SPARQL
         when RESULT_TSV
           self.class.parse_tsv_bindings(response.body, nodes)
         else
-          parse_rdf_serialization(response, options)
+          parse_rdf_serialization(response, **options)
       end
     end
 
@@ -432,9 +432,9 @@ module SPARQL
         when :uri
           RDF::URI.new(value['value'])
         when :literal
-          RDF::Literal.new(value['value'], :datatype => value['datatype'], :language => value['xml:lang'])
+          RDF::Literal.new(value['value'], datatype: value['datatype'], language: value['xml:lang'])
         when :'typed-literal'
-          RDF::Literal.new(value['value'], :datatype => value['datatype'])
+          RDF::Literal.new(value['value'], datatype: value['datatype'])
         else nil
       end
     end
@@ -550,7 +550,7 @@ module SPARQL
         when :literal
           lang     = value.respond_to?(:attr) ? value.attr('xml:lang') : value.attributes['xml:lang']
           datatype = value.respond_to?(:attr) ? value.attr('datatype') : value.attributes['datatype']
-          RDF::Literal.new(value.text, :language => lang, :datatype => datatype)
+          RDF::Literal.new(value.text, language: lang, datatype: datatype)
         else nil
       end
     end
@@ -559,8 +559,8 @@ module SPARQL
     # @param  [Net::HTTPSuccess] response
     # @param  [Hash{Symbol => Object}] options
     # @return [RDF::Enumerable]
-    def parse_rdf_serialization(response, options = {})
-      options = {:content_type => response.content_type} unless options[:content_type]
+    def parse_rdf_serialization(response, **options)
+      options = {content_type: response.content_type} unless options[:content_type]
       if reader = RDF::Reader.for(options)
         reader.new(response.body)
       else
@@ -750,7 +750,7 @@ module SPARQL
     # @see    http://www.w3.org/TR/sparql11-protocol/#query-via-get
     def make_get_request(query, headers = {})
       url = self.url.dup
-      url.query_values = (url.query_values || {}).merge(:query => query.to_s)
+      url.query_values = (url.query_values || {}).merge(query: query.to_s)
       set_url_default_graph url unless @options[:graph].nil?
       request = Net::HTTP::Get.new(url.request_uri, self.headers.merge(headers))
       request
