@@ -20,6 +20,6 @@ require 'sparql/client'
 
 RSpec.configure do |config|
   config.include(RDF::Spec::Matchers)
-  config.filter_run :focus => true
+  config.filter_run focus: true
   config.run_all_when_everything_filtered = true
 end
